backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
langchain>=0.0.200
langchain-openai>=0.0.2
openai>=1.3.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pydantic-settings>=2.7.0